Profiling: fix disabling profiling
The nixpkgs default is to enable library profiling, so turning off the
overlay won't disable profiling, we need to explicitly do it.

Also restructured the overlay a bit to be very careful about avoiding
rebuilds.

This fails for me with a linker error when trying to build the test
suite for `language-plutus-core` with profiling enabled, but that was
already true when profiling was enabled, so that's a separate issue.