Add stack2nix regeneration script
Also tweaked to be very similar to the nix-tools version. One difference
(which might be an improvement for the nix-tools script too) is taking
the output location as an argument, since we probably don't want to be
too prescriptive about that.

The hackage snapshot remains as an argument to the nix derivation so that we can
make a pinned regeneration script in a project's `default.nix`, since I would argue
that this is a slightly more important piece of data and should appear
with the rest of the nix configuration rather than be hidden in a script
argument.