Fix isLibrary checks
The Haskell generic builder falls back to checking `isExecutable` if
the `isLibrary` argument is not set. Tools like `cabal2nix` sometimes
indeed do not set `isLibrary`, so we need to mimic that logic, since we
don't have access to the "computed" answer.

Without this we incorrectly think that packages with a library but not
an executable (for which `cabal2nix` does not set `isLibrary`) are not
libraries, and thus don't Haddock them etc.