Allow control over default-nix for IFD on hydra
Using callCabalPlanToNix or callStackToNix often results in
timeouts and other eval time errors on hydra that are very difficult
to debug.

We need to be able to ask hydra to build only the `.nix`
files, but not `import` them.

Currently `default-nix` forces the plan to be imported by enumerating
the components it produces.

This change:

* Allows nix-tools-pkgs to be a function to make it possible
  parameterise.
* Allows nix-tools-pkgs to return hsPkgs separately so a the nix
  files can be returned independently.
* Adds nix-tools-raw to get access to nix-tools._raw bypassing the
  enumeration of components.

This change carries the small risk of breaking code that already
includes an `hsPkgs` attribute.