Respect recurseForDerivations = false
`recRecurseIntoAttrs` sets `recurseForDerivations = true` everything in
the tree until it finds a derivation, but with the `project` and
`package` attributes added recently to haskell.nix the tree can be
infinite.  For instance you can wind up with something like

native.haskellPackages.contra-tracer.project.hsPkgs.Cabal.project.
hsPkgs.Cabal.components.tests.check-tests.x86_64-linux

To fix this we have added explicit `recurseForDerivations = false`
to the attributes that point back to the `project` and `package`.

However currently `recRecurseIntoAttrs` does not respect that and
turns everything back to `recurseIntoAttrs = true`.

This change updates `recRecurseIntoAttrs` so it will stop whenever it
sees `recurseIntoAttrs = false`.