Reduce nix closure size of set-git-rev
With cardano-wallet, I noticed that build dependencies of set-git-rev
were not cached. Furthermore some of these dependencies were
python ← bzr ← nix-prefetch-scripts.

The problem is nixpkgs `callCabal2nix`, so this replaces it with
Haskell.nix.

There are other things in `haskell-nix-extra` which can be deprecated,
so I have removed them with deprecation errors traced where possible.