Set libblst `.version` to match `Cargo.toml`
An alternative would have been to set the version only in the `.pc` file
based on `Cargo.toml`. However, the derivation’s `.version` must be
accurate so that haskell.nix planning can detect it without building.

Therefore, we hardcode `.version` in the derivation, and add a check
against `Cargo.toml` to ensure it stays in sync.